Revise architecture: Structure InaraService as a child of EddiDataProviderService as it should be.
