+++ Tests/InaraTests.cs
@@ -1,4 +1,5 @@
-﻿using EddiInaraService;
+﻿using EddiDataDefinitions;
+using EddiInaraService;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Collections.Generic;
